test/web-platform-tests/to-upstream/cssom-view/HTMLElement-offsetWidth.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>HTMLElement.prototype.offsetWidth</title>
<link rel=help href="https://drafts.csswg.org/cssom-view/#dom-htmlelement-offsetwidth">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
"use strict";

test(() => {

  const el = document.createElement("p");
  assert_equals(el.offsetWidth, 0);

}, "offsetWidth should return 0 for an element disconnected from a document");
</script>
